pyproject.toml
[tool.poetry]
name = "bnn-edl"
version = "0.1.0"
description = "Bayesian Neural Networks with Evidential Deep Learning"
authors = ["John-Pierre Weideman <johnpierrew@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.10"
torch = "^2.9.1"
torchvision = "^0.24.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
pytorch-ignite = "^0.5.3"
wandb = "^0.23.1"
tqdm = "^4.67.1"
posteriors = "^0.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


